move picasa plugin into plugin/media subdirectory
