+++ README.md
@@ -28,11 +28,10 @@ Massr - Mini Wassr
 
 call back用のURLは『http://127.0.0.1:9393/auth/twitter/callback 』(開発用)、または、『http://HOGE-FUGA.herokuapp.com/auth/twitter/callback 』(heroku用)とする。
 
-* Googleアカウント用意
-画像のアップロードはPicasaを利用するため、Googleアカウントが必要です。
+* Googleアカウント用意[オプション]
+画像のアップロードにPicasaを利用する場合、Googleアカウントが必要です。
 IDとパスワードでログインするため、利用するアカウントで二段階認証を設定している場合はMassr用のパスワードの発行、そうでない場合は「安全性の低いアプリのアクセス」を有効にする必要があります。
 設定は https://www.google.com/settings/security で変更できます。
-
 
 ### 開発環境(development)で実行方法
 
@@ -96,7 +95,7 @@ $ bundle exec rackup --port 9393
 ```
 
 developmentでの初回起動時にはTwitterのAPI情報、Gmailのアカウント情報を設定するようviが起動します。(上記でEDITORに指定したエディタ)
-TwitterのAPI情報はユーザ認証に、Gmailのアカウント情報は画像のアップロード (Picasa) に使用します。
+TwitterのAPI情報はユーザ認証に、Gmailのアカウント情報はメールの送信に使用します。
 それぞれ ~/.pit/ 以下にファイルが作成されますので、上手く動作しないときはこの中のファイルを編集するか、一度削除して起動し直してください。
 
 http://127.0.0.1:9393 へ接続し、動作確認します。
@@ -134,10 +133,19 @@ $ heroku logs -t
 
 ### 画像投稿を有効化する方法
 
-Massrでは画像投稿にPicasaウェブアルバムを利用しています。
+画像投稿にPicasaウェブアルバムを利用する場合は、設定が必要です。
 PicasaウェブアルバムではGoogle+と連携することで、2048px x 2048px以下の画像が容量無制限でアップロード可能となります。
 
-以下の設定をすることで、画像投稿機能が有効になります。
+yamlファイルに以下の設定をすることで、Picasaを使った画像投稿機能が有効になります:
+
+```
+"plugin": {
+  "media/picasa": {
+  }
+}
+```
+
+また、Herokuの環境変数に以下の設定が必要です:
 
 ```sh
 $ heroku config:add \
