+++ app.json
@@ -8,8 +8,6 @@
   "env": {
     "TWITTER_CONSUMER_ID": "",
     "TWITTER_CONSUMER_SECRET": "",
-    "PICASA_ID": "",
-    "PICASA_PASS": "",
     "TZ": "Asia/Tokyo",
     "LANG": "en_US.UTF-8",
     "RACK_ENV": "production"
